Create Lucky Number.java
A positive integer 
�
X is called lucky if it has an even power of 
2
2 in its prime factorization. More formally, let 
�
p be the largest integer such that 
�
X is divisible by 
2
�
2 
p
 . Then 
�
X is a lucky number if and only if 
�
p is divisible by 
2
2.

You are given a positive integer 
�
N. Find if it is a lucky number.

Input Format
The first line of the input contains a single integer 
�
T - the number of test cases. The description of 
�
T test cases follows.

The first line of each test case contains a single integer 
�
N.

Output Format
For each test case, print a single line containing one integer. That integer should be 
1
1 if 
�
N is a lucky number and 
0
0 otherwise