Create Taxi Cost.java
If you travel to work by taxi, it costs you 
�
X dollars. If you travel by foot it is, of course, free of charge. You want to minimize the cost of travel, but you don’t want to walk when it’s raining or it has rained the day before. Given the weather forecast for the next 
�
N days (
1
1 if the rain is falling, 
0
0 otherwise), calculate the cost of travel.

Input Format
The first line of the input contains a single integer 
�
T denoting the number of test cases. The description of 
�
T test cases follows.
The first line of each test case contains 
2
2 integers 
�
N - number of days and 
�
X - cost for taxi
The second line contains 
�
N space-separated integers 
�
1
,
�
2
…
�
�
D 
1
​
 ,D 
2
​
 …D 
n
​
 
Output Format
For each test case, output a single line containing one integer - Total cost of the travel.