Create Find the Middle.java
You are given the head of a singly linked list 
�
A of length 
�
N. The values in the list are 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  respectively. You need to find the value of the middle element of the linked list.

The middle element of a linked list of length 
�
N is the 
(
⌊
�
2
⌋
+
1
)
(⌊ 
2
N
​
 ⌋+1)-th element from the head of the list.

Input Format
The first line of the input contains a single integer 
�
T - the number of test cases. The description of 
�
T test cases follows.

The first line of each test case contains a single integer 
�
N.

The second line of each test case contains 
�
N space-separated integers 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .

Note:

For Java language, you need to:
Complete the function in the submit solution tab:

int getMiddleElement(Node head){...}
 
 

For C++ language, you need to:
Complete the function in the submit solution tab:

int getMiddleElement(struct Node* head){...}
 
 

For Python language, you need to:
Complete the function in the submit solution tab:

def getMiddleElement(head):
Output Format
For each test case, the function you complete should return the value of the middle element of the list.