Create Remove Duplicates.java
Chef gave you the 
ℎ
�
�
�
head of a sorted linked list which may contain duplicate values. You've to remove all the duplicate values and return the sorted linked list such that it does not contain any duplicates.

Input Format
First-line will contain 
�
T, the number of test cases. Then the test cases follow.
Each test case contains two lines of input.
The first line of every test case contains an integer 
�
N - the length of array.
The second line of every test case contains 
�
N integers - 
�
1
,
�
2
,
.
.
,
�
�
A 
1
​
 ,A 
2
​
 ,..,A 
N
​
  denoting the integers in the array.
It is guaranteed that the array is sorted
You don't need to read or print anything. Just complete the function removeDuplicates() which takes the head of the linked list as input.
Output Format
Return the head of the sorted linked list without duplicates.