Create  Rotate the List Difficulty Rating:  1300.java
You are given the head of a singly linked list 
�
A of length 
�
N. The values in the list are 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  respectively. You are also given a non-negative integer 
�
R. You need to rotate the list 
�
R places to the right.

A single rotation to the right is an operation in which the last element of the list is moved to the first place in the list, while all the other elements are moved one place to the right.

Input Format
The first line of the input contains a single integer 
�
T - the number of test cases. The description of 
�
T test cases follows.

The first line of each test case contains two space-separated integers 
�
N and 
�
R.

The second line of each test case contains 
�
N space-separated integers 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .

Note:

For Java language, you need to:
Complete the function in the submit solution tab:

Node rotateRight(Node head, int R){...}
 
 

For C++ language, you need to:
Complete the function in the submit solution tab:

Node* rotateRight(struct Node* head, int R){...}
 
 

For Python language, you need to:
Complete the function in the submit solution tab:

def rotateRight(head, R):
Output Format
For each test case, the function you complete should return the head of the list which has been rotated to the right exactly 
�
R times.