Create Reverse the Segment Difficulty Rating:  1500.java
You are given the head of a singly linked list 
�
A of length 
�
N. The values in the list are 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  respectively. You are also given two integers 
�
L and 
�
R. You need to reverse the nodes of the list from position 
�
L to position 
�
R.

Input Format
The first line of the input contains a single integer 
�
T - the number of test cases. The description of 
�
T test cases follows.

The first line of each test case contains three space-separated integers 
�
N, 
�
L and 
�
R.

The second line of each test case contains 
�
N space-separated integers 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .

Note:

For Java language, you need to:
Complete the function in the submit solution tab:

Node reverseSegment(Node head, int L, int R){...}
 
 

For C++ language, you need to:
Complete the function in the submit solution tab:

Node* reverseSegment(struct Node* head, int L, int R){...}
 
 

For Python language, you need to:
Complete the function in the submit solution tab:

def reverseSegment(head, L, R):
Output Format
For each test case, the function you complete should return the head of the list in which the nodes from the appropriate segment are reversed.