Create  Duplicate Removal Difficulty Rating:  1200.java
The Chef is celebrating his birthday! His best friend gifted him a sorted singly linked list. But there is a catch - Chef hates duplicates, so he wants to remove every element that occurs more than once (he wants to remove all of the occurrences). However, he is very busy with his birthday party, he asks you to help him.

Complete the function "duplicateRemoval" in the code snippet that takes a single argument: a head of the sorted singly linked list. The function should return the head of the singly linked list that is the modification of the original one, but the duplicates are completely erased. The function will be called multiple times.

Input Format
First line will contain 
�
T, number of testcases. Then the testcases follow.
The first line of each test case contains an integer 
�
N - representing the length of the linked list
The second line of each test case contains 
�
N space-separated integers - representing the elements of the list
Note: You need to complete the functions in the submit solution tab:

Node* duplicateRemoval(Node* head){...}

Output Format
For each testcase, output will be in a single line containing a list passed by the function duplicateRemoval.