Create  Crits in a Linked List! Difficulty Rating:  1000.java
Given the head of a linked list, Find the number of critical points. (The starting and end are not considered critical points).

Local minima or maxima are called critical points.

A Node is called a local minima if both next and previous elements are greater than the current element.

A Node is called a local maxima if both next and previous elements are smaller than the current element.

Input Format
The first line contains 
�
N, the number of elements in the linked list.
The second line contains a number 
�
m which denotes where the loop will be made from the tail of the linked list. 
0
≤
�
<
�
0≤m<n (i.e the first node is called 
0
�
ℎ
0 
th
  node).
(Naturally the Linked list has connections from 
�
�
ℎ
i 
th
  node to 
�
+
1
�
ℎ
i+1 
th
  node).
(Note. Node values will be zero)
 
 

For C++ language, you need to:
Complete the function in the submit solution tab:

int solve(Node* head){}
 
 

Output Format
-1 if a loop doesnt exist, else the length of the loop.