Create  Strange List Difficulty Rating:  1600.java
Alice has a strange singly-linked list. A usual singly linked-list or a not-strange linked list consist of nodes pointing to exactly one node ( except for the the tail node which points to null ). However, in case of a strange singly-linked list, instead of pointing towards a single linked-list node, some nodes of the linked-list are pointing towards two nodes! Each node in the strange linked-list has two pointers named :

next
child
The “child” pointer of a node may again point towards a strange singly linked list.

Alice wants to restructure this linked-list into a not-strange linked list. While doing so, she must follow the following rules.

Let cur be a node in restructured linked-list, then

cur.next must occur after cur in the restructured linked-list.
All the nodes ( if any ) which are part of cur.child linked list must occur after the node cur and before the node cur.next in the restructured linked list.
Let cur be a node in the restructured linked-list, then cur.child must be “null”.
NOTE : Refer to sample case for a visual depiction.

Input Format
The first line of the input contains a single integer T - the number of test cases. The description of T test cases follows.

The first line of each test case contains a single integer 
�
N.

The following 
�
−
1
N−1 lines contain three space separated integers 
�
,
�
u,v and 
�
�
�
�
type. If the value of 
�
�
�
�
type is equal to 0, then the next pointer of node with value equal to 
�
u points towards the node with value equal to 
�
v. If the value of type is equal to 1, then the child pointer of node with value equal to 
�
u points towards the node with value equal to 
�
v.

For C++ language, you need to:

Complete the function in the submit solution tab:

bool flatten(Node* head){ . . . }

Output Format
For each test case, the function you complete should return the restructured linked-list.