Create  Train Loop.java
You are given a linked-list with 
�
N nodes. It might be possible that the tail of the linked-list points to the head of the linked-list; If this is the case, then we call the linked-list to be circular.

Each node (except the tail node) points to exactly one node, denoted by the pointer next. Each node has a value associated with itself, denoted by the field val.

Input Format
The first line of the input contains a single integer 
�
T - the number of test cases. The description of T test cases follows.
The first line of each test case contains two space separated integers N and M.
The following 
�
M lines contain two space separated integers u and v, denoting that there is a connection directed from compartment u to compartment v.
For C++ language, you need to: Complete the function in the submit solution tab:
bool isCircular(Node* engine){ . . . }

Output Format
For each test case, the function you complete should return a boolean. “true” if the train forms a loop, “false” otherwise.