Create Good Numbers.java
Given four positive integers 
�
N, 
�
a, 
�
b and 
�
c. A number is called good if it is positive and is divisible by either 
�
a or 
�
b or 
�
c. If all the good numbers are listed in ascending order, find the 
�
�
ℎ
N 
th
  good number.

Input Format
First line will contain 
�
T, number of testcases. Then the testcases follow.
Each testcase contains of a single line of input, four integers 
�
,
�
,
�
,
�
N,a,b,c.
Output Format
For each testcase, output in a single line the 
�
�
ℎ
N 
th
  good number.

Constraints
1
≤
�
≤
1000
1≤T≤1000
1
≤
�
≤
1
0
9
1≤N≤10 
9
 
1
≤
�
,
�
,
�
≤
1
0
3
1≤a,b,c≤10 
3
 
Subtasks
15 points : 
1
≤
�
,
�
,
�
,
�
≤
10
1≤a,b,c,N≤10
85 points : original constraints