Create  Increasing-Decreasing Sequence.java
You are given a sequence 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 . You need to find the maximum length of a subsequence of 
�
A whose elements first strictly increase and then strictly decrease in value. Note that subsequences that are only strictly increasing and subsequences that are only strictly decreasing are considered as well.

A subsequence of a given sequence is a sequence that can be derived from the given sequence by deleting some or no elements without changing the order of the remaining elements.

Input Format
The first line of the input contains a single integer 
�
T denoting the number of test cases. The description of 
�
T test cases follows.

The first line of each test case contains an integer 
�
N.

The second line contains 
�
N space-separated integers 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .

Output Format
For each test case, print a single line containing one integer ― the maximum length of a subsequence of the sequence 
�
A which is first strictly increasing and then strictly increasing.
Constraints
1
≤
�
≤
100
1≤T≤100
1
≤
�
≤
2
⋅
1
0
5
1≤N≤2⋅10 
5
 
the sum of 
�
N over all test cases does not exceed 
2
⋅
1
0
5
2⋅10 
5
 
1
≤
�
�
≤
1
0
9
1≤A 
i
​
 ≤10 
9
  for each valid 
�
i
Subtasks
Subtask #1 (30 points):

�
≤
2000
N≤2000
the sum of 
�
N over all test cases does not exceed 
4000
4000
Subtask #2 (70 points): original constraints

Sample 1:
Input
Output
4
9
1 2 3 1 4 1 3 2 1
5
2 4 3 6 3
4
1 2 3 4
5
9 9 9 9 9
7
4
4
1
Explanation:
Example case 1: The longest suitable subsequence is 
[
1
,
2
,
3
,
4
,
3
,
2
,
1
]
[1,2,3,4,3,2,1].

Example case 2: The longest suitable subsequence is 
[
2
,
4
,
6
,
3
]
[2,4,6,3].

Example case 3: Although it is only strictly increasing, the longest suitable subsequence is 
[
1
,
2
,
3
,
4
]
[1,2,3,4].

Example case 4: The longest suitable subsequence is 
[
9
]
[9].